+++ tests/test_fusionfisher.py
@@ -24,8 +24,13 @@ from .data_features import *
     'misannotation_both_nonpseudo',
     'misannotation_one_pseudo',
     'idempotence',
+    'consecutive_fusions',
 ])
 def test_fusionfisher(gene_list):
+
+    if gene_list == 'consecutive_fusions':
+        pytest.skip("incomplete test case")
+
     source_genomes = {
         'whole_gene_fusion':'1-0006',
         'misannotation_false_delayed_stop':'SEA17020030',
@@ -34,6 +39,7 @@ def test_fusionfisher(gene_list):
         'misannotation_both_nonpseudo': 'AZ20',
         'misannotation_one_pseudo': 'AZ20',
         'idempotence': 'AZ20',
+        'consecutive_fusions': 'AZ20_reannotatedref',
     }
     # create a dummy feature dictionary for the test case that is not in focus in the current invocation to avoid a KeyError when loading all expected inputs and results
     source_features = defaultdict(lambda :defaultdict(dict))
@@ -67,6 +73,12 @@ def test_fusionfisher(gene_list):
             deepcopy(source_features['ECOLIN_24700']['ratt']),
             deepcopy(source_features['ECOLIN_12095']['prokka']),
         ],
+        'consecutive_fusions': [
+            source_features['ECOLIN_26716']['ratt'],
+            source_features['ECOLIN_26715']['ratt'],
+            source_features['ECOLIN_26714']['ratt_corrected'],
+            source_features['ECOLIN_26713']['ratt'],
+        ]
     }
     ref_genome = defaultdict(lambda :'H37Rv')
     ref_genome.update({
@@ -74,6 +86,7 @@ def test_fusionfisher(gene_list):
         'misannotation_both_nonpseudo': 'nissle-hybrid',
         'misannotation_one_pseudo': 'nissle-hybrid',
         'idempotence': 'nissle-hybrid',
+        'consecutive_fusions': 'ECOLIN',
     })
     source_genome = source_genomes[gene_list]
 
@@ -89,6 +102,7 @@ def test_fusionfisher(gene_list):
             f, annomerge.ref_annotation[key_ref_gene(f.source, f.qualifiers['gene'][0])]
         )
 
+    # outputs are [outlist, fusions, rejects]
     expected = {
         'whole_gene_fusion': (
             deepcopy(inputs['whole_gene_fusion']),
@@ -145,6 +159,11 @@ def test_fusionfisher(gene_list):
             [ inputs['idempotence'][0] ],
             [ ],
         ),
+        'consecutive_fusions': (
+            deepcopy(inputs['consecutive_fusions']),
+            [],
+            [],
+        ),
     }
     if gene_list == 'idempotence':
         expected['idempotence'][0][1].qualifiers['note'] = [
@@ -181,6 +200,60 @@ def test_fusionfisher(gene_list):
         ]
         component1.location = FeatureLocation(392626, 392851, strand=1, ref='1')
         expected['whole_gene_fusion'][1].append(expected['whole_gene_fusion'][0][0])
+    elif gene_list == 'consecutive_fusions':
+        # TODO: check all these fusion component coordinates. see if they make sense
+        #
+        # NOTE**  - NISSLEORF0154 fusions occur twice. There seem to be two consecutive copies that got fused
+        #           but our practice is to not name them twice (...::NISSLEORF0154::NISSLEORF0154)
+#         [
+# 1[2651439:2658051](-)
+# Fusion type: whole
+# Components:
+# NISSLEORF0154|1[2656502:2658051](-)
+# ECOLIN_26716|NISSLEORF0143|1[2651439:2656341](-)
+# Qualifiers:
+# {'locus_tag': ['ECOLIN_26714'], 'gene': ['NISSLEORF0154::NISSLEORF0143'], 'note': ['Upstream gene ECOLIN_26714|NISSLEORF0154 conjoins with this one.']}
+# ,
+# 1[2651439:2659227](-)
+# Fusion type: whole
+# Components:
+# NISSLEORF0154|1[2656502:2659227](-)
+# ECOLIN_26714|NISSLEORF0154::NISSLEORF0143|1[2651439:2658051](-)
+# Qualifiers:
+# {'locus_tag': ['ECOLIN_26714'], 'gene': ['NISSLEORF0154::NISSLEORF0143'], 'pseudo': [''], 'note': ['Hybran/Pseudoscan:description:Locus does not have reference-corresponding start | No internal stop codons and ends with a valid stop codon | Locus has valid reading frame | Locus has a valid alternative start site', 'Hybran/Pseudoscan:evidence:no_rcc', 'Hybran/Pseudoscan:barcode:D31;VS1;VE1;RCS0;RCE1;BOK.', 'Upstream gene ECOLIN_26713|NISSLEORF0188 conjoins with this one.']}
+# ,
+# 1[2651439:2659977](-)
+# Fusion type: whole
+# Components:
+# NISSLEORF0188|1[2659130:2659977](-)
+# ECOLIN_26714|NISSLEORF0154::NISSLEORF0143|1[2651439:2659227](-)
+# Qualifiers:
+# {'locus_tag': ['ECOLIN_26713'], 'gene': ['NISSLEORF0188::NISSLEORF0154::NISSLEORF0143']}
+# ]
+        expected['consecutive_fusions'][1] = expected['consecutive_fusions'][0][1:]
+        # first fusion
+        expected['consecutive_fusions'][1][0].qualifiers['note'] = [
+            'Upstream gene ECOLIN_26714|NISSLEORF0154 conjoins with this one.',
+        ]
+        expected['consecutive_fusions'][1][0].fusion_type = "whole"
+        componentX = deepcopy(expected['consecutive_fusions'][1][0])
+        (componentX.location._start, componentX.location._end)  = (2656502, 2658051)
+        expected['consecutive_fusions'][1][0].components = [
+            componentX,
+            expected['consecutive_fusions'][0][0],
+        ]
+        # second fusion
+        expected['consecutive_fusions'][1][1].qualifiers['note'] = [
+            'Upstream gene ECOLIN_26713|NISSLEORF0188 conjoins with this one.',
+        ]
+        expected['consecutive_fusions'][1][1].fusion_type = "whole"
+        componentY = deepcopy(expected['consecutive_fusions'][1][1])
+        (componentY.location._start, componentY.location._end)  = (2656502, 2658051)
+        expected['consecutive_fusions'][1][1].components = [
+            expected['consecutive_fusions'][0][0],
+            #....
+        ]
+        # third fusion
 
     # set the rival feature as the passing one. for these test cases, we're always looking at pairs
     if expected[gene_list][2]: