+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fixed off by 1 error when reporting the location of translational exceptions (selenocysteine proteins). (#72)
 * Lifted memory restrictions on cd-hit (#74)
 * Fixed crash that occurs when only one gene component of a fusion doesn't have a product field. (#75)
+* Sort features by end position when start positions are identical.
 
 ## [Version 1.8](https://gitlab.com/LPCDRP/hybran/-/tags/1.8)